# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000"
info:
  title: Fantastic Coffee (decaffeinated)
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
security:
  - bearerAuth: []
components:   
  securitySchemes:
    bearerAuth:     # arbitrary name for the security scheme
      description: "arbitrary name for the security scheme"       
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      headers:
        WWW_Authenticate:
          schema:
            type: string
  schemas:
    ID: 
      title: General ID format 
      description: "Specifies the format of a unique identifier (UUID)."
      type: string 
      pattern: '^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$'
      minLength: 36
      maxLength: 36
      example: 3e7b82c1-8f66-4f54-9e57-7baa8a0ea8e3
               
               # 1a2b3c4d-5e6f-7a8b-9c0d-1e2f3a4b5c6d
      readOnly: true
    Comment:
      title: Comment of a post
      description: |-
        comment provided by and user to an images of another user
      type: object 
      properties:
        # identifier of the post to which the comment belongs 
        post_id:
          $ref: "#/components/schemas/ID"

        # identifier of the comment/like in a post, given that we cann have many of the same user
        comment_id:
          $ref: "#/components/schemas/ID"

        # id of the user who does the comment/like, this will help us identify from whom the comment/like is
        user_id: 
          $ref: "#/components/schemas/ID"
          
        # the comment per se
        body:
          type: string
          description: "String representing a comment of a user"
          pattern: "^[a-zA-Z0-9]+$"
          minLength: 1
          maxLength: 2204
...
paths:
...
  /comments/:
    post:
      security: 
        - bearerAuth: []
      tags:
        - Post
      operationId: commentPhoto
      summary: add a comment to a post
      description: |- 
        an user want to add a comment to a post of him self of another user post 
          
      requestBody:
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Comment"}
      responses:
        "201":
          description: Comment successfully created 
          # we need a response body given that we are creating a new resources
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Comment"}
              
        '400':
          description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        
        '500':
          description: Internal server error.
  
  # UNCOMMENT PHOTO
  # here we want to remove a comment but with a simple constraint that only the "owner" of the comment can delete it's own comment  
  /comments/{commentid}:
    delete:
      security: 
        - bearerAuth: []
      tags: 
        - Posts
      operationId: uncommentPhoto
      summary: Remove a comment from a post
      description: |-
        An user want to remove a comment done in a post 
      parameters:
        - name: commentid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true 
          
      responses:
        "204":
          description: |- 
            specify if the comment in the user post was correctly deletated 
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
            
        "404":
          description: Resource not found
          
        "500":
          description: Internal server error 
          
  # SET MY USER NAME
  # an user want to change it's username
...
  /users/{userid}/following:
    post:
      security:
        - bearerAuth: []
      tags:
        - Users
      operationId: followUser
      summary: Start follow another user
      description: |-
        An user wants to start to follow another user 
      
      parameters:
        - name: userid
          in: path
          description: ID of the user who want to follow another user
          schema:
            $ref: "#/components/schemas/ID" 
          required: true 
      
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: "ID of the user that we want to follow, so ID of the user to be followed"
              properties:
                followedID:
                  # ID of the user that we want to follow, so ID of the user to be followed.
                  $ref: "#/components/schemas/ID"
                followingUsername:
                  type: string
                  pattern: '^.*?$'
                  description: Username of the user to be followed.
                  example: 'el_jefe'
                  minLength: 3 
                  maxLength: 16 
      
      responses:
        "201":
          description: user followed successfully 
          content:
            application/json:
              schema:
                type: object
                description: "ID of the user that we want to follow, so ID of the user to be followed."
                properties:
                
                  followedID:
                    # ID of the user that we want to follow, so ID of the user to be followed.
                    $ref: "#/components/schemas/ID"
               
                  followingUsername:
                    type: string
                    pattern: "ˆ.*?$"
                    description: Username of the user to be followed.
                    example: 'el_jefe'
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        
        "404":
          description: Resource not found
        
        "500":
          description: Internal server error 
        
  # UNFOLLOW USER
  # user want ot unfollow and user that start follow early
  /users/{userid}/following/{followedid}:
    delete:
      security: 
        - bearerAuth: []
      tags:
        - Users
      operationId: unfollowUser
      summary: user unfollow an user
      description: |-
        User want to unfollow  another user that follow in past
      parameters:
        - name: userid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
        
        - name: followedid
          in: path
          schema:
            $ref: "#/components/schemas/ID"
          required: true
          
      responses:
        "204":
          description: |- 
            specify if the followed user is not more in the collection of folliwing users 
            
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        "404":
          description: Resource not found
          
        "500":
          description: Internal server error 
          
  # BAND USER
  # here we create a collection of muted user for a particular userid    
  /users/{userid}/muted/:    
  # /users/{userid}/muted/{muteduserid}:
    post:
      security: 
        - bearerAuth: []
      tags:
        - Users
      operationId: banUser 
      summary: an user band another user
      description: |-
        Mutes another user, preventing them from interacting with the specified user.
      parameters:
        - name: userid
          in: path
          description: ID of the user who wants to mute another user.
          required: true
          schema:
            $ref: "#/components/schemas/ID"
            
      requestBody:
        content:
          application/json:
            schema:
              type: object
              description: "ID of the user to be muted."
              properties:
                mutedUserId:
                  # ID of the user to be muted.
                  $ref: "#/components/schemas/ID"
      responses:
        '201':
          description: User muted successfully.
          content:
            application/json:
              schema:
                type: object
                description: "the userid of the muted person"
                properties:
                  mutedUserId:
                    # ID of the user to be muted.
                    $ref: "#/components/schemas/ID"

        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found.
        '500':
          description: Internal server error.

  # UNBAND USER
  # here we want to unmute an user from the muted collection 
  /users/{userid}/muted/{mutedUserid}:
    delete:
      security:
        - bearerAuth: []
      tags:   
        - Users
      operationId: unbanUser
      summary: an user unban another user 
      description: |-
        Allows a user to unmute another user, allowing them to interact again.
      parameters:
        - name: userid
          in: path
          description: ID of the user who wants to unmute another user.
          required: true
          schema:
            $ref: "#/components/schemas/ID"
        - name: mutedUserid
          in: path
          description: ID of the user to be unmuted.
          required: true
          schema:
            $ref: "#/components/schemas/ID"
      responses:
        '204':
          description: User unmuted successfully.
        '400':
            description: Bad request. Invalid user data.
            
        '401':
            $ref: '#/components/responses/UnauthorizedError'
        '404':
          description: User not found or not muted.
        '500':
          description: Internal server error.  
          
  # GET USER PROFILE 
  # we use the path , without including the query parameters directly in the path, 
  #it should be done when an user makes a request to search an user e.g  /users/{userid}/profile?username=desiredUsername
...